# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfileStatus = typeof ProfileStatus[keyof typeof ProfileStatus];


export const ProfileStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface Profile {
  id: string;
  displayName: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  country: string;
  languagesSpoken: string[];
  languagesLearning: string[];
  interests: string[];
  bio: string;
  status: ProfileStatus;
  lastActiveAt?: string;
}

export type ProfileUpdateMatchingMode = typeof ProfileUpdateMatchingMode[keyof typeof ProfileUpdateMatchingMode];


export const ProfileUpdateMatchingMode = {
  random: 'random',
  language_exchange: 'language_exchange',
  shared_interests: 'shared_interests',
  culture_exchange: 'culture_exchange',
} as const;

export interface ProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 40
     */
  displayName?: string;
  /** @maxLength 180 */
  bio?: string;
  country?: string;
  languagesSpoken?: string[];
  languagesLearning?: string[];
  interests?: string[];
  preferredCountries?: string[];
  preferredLanguages?: string[];
  matchingMode?: ProfileUpdateMatchingMode;
}

export type AppStateStats = {
  conversations: number;
  countries: number;
  minutes: number;
};

/**
 * @nullable
 */
export type AppStateActiveMatch = {
  matchId?: string;
  partner?: Profile;
  prompt?: string;
  connectedAt?: string;
} | null;

export interface Notification {
  id: string;
  title: string;
  message: string;
  createdAt: string;
  read: boolean;
}

export interface AppState {
  profile: Profile;
  stats: AppStateStats;
  /** @nullable */
  activeMatch: AppStateActiveMatch;
  notifications: Notification[];
}

export interface CountryCard {
  code: string;
  name: string;
  flag: string;
  conversationCount: number;
  description: string;
}

export interface DiscoverResponse {
  headline: string;
  countries: CountryCard[];
  interests: string[];
}

export type MatchPreferencesMode = typeof MatchPreferencesMode[keyof typeof MatchPreferencesMode];


export const MatchPreferencesMode = {
  random: 'random',
  language_exchange: 'language_exchange',
  shared_interests: 'shared_interests',
  culture_exchange: 'culture_exchange',
} as const;

export interface MatchPreferences {
  countries: string[];
  languages: string[];
  interests: string[];
  mode: MatchPreferencesMode;
}

export type MatchStateStatus = typeof MatchStateStatus[keyof typeof MatchStateStatus];


export const MatchStateStatus = {
  idle: 'idle',
  searching: 'searching',
  found: 'found',
  connected: 'connected',
} as const;

export interface MatchState {
  status: MatchStateStatus;
  /** @nullable */
  matchId?: string | null;
  partner?: Profile | null;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  queuePosition?: number | null;
}

export interface Message {
  id: string;
  senderId: string;
  message: string;
  createdAt: string;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
}

export interface HistoryItem {
  id: string;
  date: string;
  country: string;
  durationMinutes: number;
  interests: string[];
  partnerName: string;
  hidden?: boolean;
}

export type ReportInputReason = typeof ReportInputReason[keyof typeof ReportInputReason];


export const ReportInputReason = {
  harassment: 'harassment',
  spam: 'spam',
  scam: 'scam',
  hate_speech: 'hate_speech',
  threats: 'threats',
  inappropriate_behavior: 'inappropriate_behavior',
  sexual_content: 'sexual_content',
  underage_concern: 'underage_concern',
  other: 'other',
} as const;

export interface ReportInput {
  reason: ReportInputReason;
  /** @maxLength 1000 */
  description?: string;
  blockUser?: boolean;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  pending: 'pending',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

export interface Report {
  id: string;
  reason: string;
  status: ReportStatus;
  createdAt: string;
}

export interface BlockInput {
  userId: string;
}

export interface BlockedUser {
  userId: string;
  displayName: string;
  country: string;
  blockedAt: string;
}