# lib/api-zod/src/generated/types/appState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { AppStateActiveMatch } from './appStateActiveMatch';
import type { AppStateStats } from './appStateStats';
import type { Notification } from './notification';
import type { Profile } from './profile';

export interface AppState {
  profile: Profile;
  stats: AppStateStats;
  /** @nullable */
  activeMatch: AppStateActiveMatch;
  notifications: Notification[];
}