# lib/api-zod/src/generated/types/blockedUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockedUser {
  userId: string;
  displayName: string;
  country: string;
  blockedAt: string;
}