# lib/api-zod/src/generated/types/countryCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */

export interface CountryCard {
  code: string;
  name: string;
  flag: string;
  conversationCount: number;
  description: string;
}