# lib/api-zod/src/generated/types/matchPreferences.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchPreferencesMode } from './matchPreferencesMode';

export interface MatchPreferences {
  countries: string[];
  languages: string[];
  interests: string[];
  mode: MatchPreferencesMode;
}