# lib/api-zod/src/generated/types/matchState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchStateStatus } from './matchStateStatus';
import type { Profile } from './profile';

export interface MatchState {
  status: MatchStateStatus;
  /** @nullable */
  matchId?: string | null;
  partner?: Profile | null;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  queuePosition?: number | null;
}