# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileStatus } from './profileStatus';

export interface Profile {
  id: string;
  displayName: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  country: string;
  languagesSpoken: string[];
  languagesLearning: string[];
  interests: string[];
  bio: string;
  status: ProfileStatus;
  lastActiveAt?: string;
}