# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileUpdateMatchingMode } from './profileUpdateMatchingMode';

export interface ProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 40
     */
  displayName?: string;
  /** @maxLength 180 */
  bio?: string;
  country?: string;
  languagesSpoken?: string[];
  languagesLearning?: string[];
  interests?: string[];
  preferredCountries?: string[];
  preferredLanguages?: string[];
  matchingMode?: ProfileUpdateMatchingMode;
}