# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportStatus } from './reportStatus';

export interface Report {
  id: string;
  reason: string;
  status: ReportStatus;
  createdAt: string;
}