# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * GlobalTalk API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputReason } from './reportInputReason';

export interface ReportInput {
  reason: ReportInputReason;
  /** @maxLength 1000 */
  description?: string;
  blockUser?: boolean;
}